Add all needed services and ends db relationships

+++ contact_book_back_end/src/controllers/UserController.js
@@ -13,67 +13,119 @@ function generateToken(params = {}) {
 module.exports = {
   //############## GET ##############
   async getAllUsers(req, res) {
-    const users = await User.findAll();
+    try {
+      const users = await User.findAll();
 
-    if (users == "" || users == null)
-      return res.status(200).send({message: "No registered users!"});
+      if (users == "" || users == null)
+        return res.status(200).send({message: "No registered users!"});
+      
+      return res.status(200).send({users});
     
-    return res.status(200).send({users});
+    } catch (e) {
+      return res.status(400).json({
+        status: false,  
+        message: 'Error getting all users!',
+        error: e
+      });
+    }
+  },
+
+  async getUser(req, res) {
+    const {user_id} = req.params;
+    
+    try {   
+      const user = await User.findByPk(user_id);
+
+      if (!user)
+        return res.status(400).send({status: false, message: 'User not found!'});
+
+      // Dont show the password in the response
+      user.password = undefined
+
+      return res.status(200).send(user);
+    
+    } catch (e) {
+      return res.status(400).json({
+        status: false,  
+        message: 'Error getting user!',
+        error: e
+      });
+    }
   },
 
   //############## POST ##############
   async login(req, res) {
     const {password, email} = req.body;
-    const user = await User.findOne({where: {email}});
-
-    if (!user) {
-      return res.status(400).send({
-        status: false,
-        message: 'Incorrect email or password!',
-        user: {}
-      });
-    }
-
-    // Password's verify after hash application
-    if (!bcrypt.compareSync(password, user.password)) {
-      return res.status(400).send({
-        status: false,
-        message: 'Incorrect email or password!',
-        user: {}
-      });
-    }
-
-    // Dont show the password in the response
-    user.password = undefined
     
-    const token = generateToken({id: user.id});
+    try {
+      const user = await User.findOne({where: {email}});
 
-    return res.status(200).send({
-      status: true,
-      message: "User successfully logged in!",
-      user, 
-      token
-    });
+      if (!user) {
+        return res.status(400).send({
+          status: false,
+          message: 'Incorrect email or password!',
+          user: {}
+        });
+      }
+
+      // Password's verify after hash application
+      if (!bcrypt.compareSync(password, user.password)) {
+        return res.status(400).send({
+          status: false,
+          message: 'Incorrect email or password!',
+          user: {}
+        });
+      }
+
+      // Dont show the password in the response
+      user.password = undefined
+      
+      const token = generateToken({id: user.id});
+
+      return res.status(200).send({
+        status: true,
+        message: "User successfully logged in!",
+        user, 
+        token
+      });
+    
+    } catch (e) {
+      return res.status(400).json({
+        status: false,  
+        message: 'Login error!',
+        error: e
+      });
+    }
   },
 
   async createUser(req, res) {
     const {name, password, email} = req.body;
-    const userAlreadyRegistered = await User.findOne({where: {email}});
 
-    // UniqueKey verification
-    if(userAlreadyRegistered)
-      return res.status(400).send({status: false, message: "E-mail already registered!"});
+    try {
+      const named_user = await User.findOne({where: {email}});
 
-    const user = await User.create({name, password, email});
+      // Unique verification
+      if(named_user)
+        return res.status(400).send({status: false, message: "E-mail already registered!"});
 
-    // Dont show the password in the response
-    user.password = undefined
+      const user = await User.create({name, password, email});
 
-    return res.status(200).send({
-      status: true,
-      message: 'Successfully registered user!',
-      user
-    });
+      // Dont show the password in the response
+      user.password = undefined
+
+      return res.status(200).send({
+        status: true,
+        message: 'Successfully registered user!',
+        user
+      });
+    
+    } catch (e) {
+      return res.status(400).json({
+        status: false,  
+        message: 'Error creating user!',
+        error: e
+      });
+    }
   },
 
   //############## PUT ##############
@@ -81,23 +133,51 @@ module.exports = {
     const {name, password, email} = req.body;
     const {user_id} = req.params;
 
-    await User.update({name, password, email}, {where: {id: user_id}});
+    try {
+      await User.update({name, password, email}, {where: {id: user_id}});
 
-    return res.status(200).send({
-      status: true,
-      message: "User updated successfully!",
-    });
+      return res.status(200).send({
+        status: true,
+        message: "User updated successfully!",
+      });
+    
+    } catch (e) {
+      return res.status(400).json({
+        status: false,  
+        message: 'Error updating user!',
+        error: e
+      });
+    }
   },
 
   //############## DELETE ##############
   async deleteUser(req, res) {
     const {user_id} = req.params;
 
-    await User.destroy({where: {id: user_id}});
+    try {
+      const user = await User.findByPk(user_id);
 
-    return res.status(200).send({
-      status: true,
-      message: "User deleted successfully!",
-    });
+      if (user) {
+        await User.destroy({where: {id: user_id}});
+
+        return res.status(200).json({
+          status: true,
+          message: "User successfully deleted!",
+        });
+
+      } else {
+        return res.status(400).json({
+          status: false,
+          message: 'User not found!'
+        });
+      }
+
+    } catch (e) {
+      return res.status(400).json({
+        status: false,  
+        message: 'Error deleting user!',
+        error: e
+      });
+    }
   }
 };
